Utility: Uninit slice from MaybeUninit slice
This also enables the common case for a `FixedVec` to be initialized
without any casts on the underlying memory resource. Note that this will
–for now– only be ergonomic when the type contained in the vector is
`Copy` since large `MaybeUninit`-array creation also requires `Copy`
bound. This may change soon, when `const fn` are treated like `Copy`
types for the sake of array initialization and `MaybeUninit::uninit` can
thus be freely used.