Fix allocation of zero-length slices
Note that we do not add a get_slice class for raw layouts. This would
commit to a particular representation for such layouts and I'm not sure
if it is better to unbound the type parameter on Allocation as ?Sized or
a different way. The main interface is exposed only due to prior art of
vectors and extend_from_slice.

Slices are special since their strictly defined layout will allows us to
build them incrementally while this is not true for other types. But
this does not require a public interface and I would rather the `get`
interface be used sparingly. Maybe it was a mistake to even expose it.
The trouble is that the semantics of generic unsized pointers are not
entirely clear. In particular, are there types like c_str that already
require a validity invariant on the raw pointee? Maybe we can add a
trait to alloc-traits in the future to answer this instead.