Expose default implementable methods as functions
This allows the use of them as a fallback option in implementations.
Such behaviour is already present in the std version of the trait where
grow_in_place is a specialized operation that is tried before the
forcible new allocation. But when one wants to perform this in a trait
implementation, then the simple implementation needs to be repeated.