Unify allocator methods into MemBump
Currently the allocation is logically replicated for each of the generic
instances of `Bump<T>`. The main observation is that we should treat
each as a simple byte buffer though. So, instead we introduce Deref such
that each can be treated as the _same_ underlying (dynamically sized)
type that contains a slice of MaybeUninit bytes behind an UnsafeCell.

This should help with compiled code size as well as there are no two
monomorphizations of the same allocator methods.