Relax Uninit/UninitView for unsized types
This was previously blocked on ptr::slice_from_raw_parts_mut to avoid
the dereference when splitting off slices. Now we can store the original
typed pointer into our own representation without the need to cast. This
preserves the original tag information which in particular allows a
proper construction from an unsized pointer.