Coalesce `Coerce` and `CoercionWith` into one generic type + reduce drive bounds.
The second part of this commit may be controversial, given how it may hinder
code readability. That being said, it's a solution that involves no dependencies
(_e.g._, `::derivative`) and that reads well in the docs (mainly, when
`cfg(rustdoc)` is on, the macro does nothing).

Otherwise it defines an internal helper struct which is _generic over its fields_,
applies the `#[derive]`s to it so that the generated bounds effectively check
the fields, and finally defines the original type as a type alias forwarding the
real generics parameters to the generic field definitions.

One drawback of this approach has been to have to loosen the `#[repr(transparent)]`
annotation to a `#[repr(C)]` one; I think the benefits of having general
`Eq`, `Hash`, `Copy` bounds for the `F = fn(…)` case, for instance (which is tested
in the newly added `src/tests.rs`) outweights the only drawback of `#[repr(C)]`
_vs._ `#[repr(transparent)]`: not being able to transmute function pointers using
`Coercion`s (by value) within its signature.