Add unsafe unsync::Bump::get_unchecked trapdoor
This is the first interaction where a `Level` previously observed can be
used again. The interface shows its first signs of cracks: This only
really works due to being used on an unsync allocator for which we do
not allow or get unsynchronized concurrency. Otherwise, there'd be no
way to retrieve a guarantee that the observed level matches the
retrieved allocations.

The `Bump` interface only partially returns the right `Level`, so it can
not be mapped directly. We'll need a new major version of the library.

Then we can also fix the provenance UB in the iterator-collector!