Fix stacked borrows UB for fixed_vec
The pointer `as_ptr` and `as_mut_ptr`, which was assumed to be valid for
the whole allocation, was derived from `Deref` and thus only the valid
part constructed by `from_raw_parts(_, self.len())`. Miri complained
correctly when this was used to write to the slot at `self.len()` in
`extend` etc.